stage1_cifar10/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
matplotlib>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
mlflow>=2.8.0
